+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentWithAnnotatedParameterCallback.java
@@ -38,7 +38,6 @@ abstract class React4j_ComponentWithAnnotatedParameterCallback extends Component
   }
 
   @Override
-  @Action
   protected void reportPropsChanged(@Nullable final JsPropertyMap<Object> nextProps) {
   }