+++ processor/src/test/resources/expected/com/example/arez/React4j_ComponentWithCallback.java
@@ -39,7 +39,6 @@ abstract class React4j_ComponentWithCallback extends ComponentWithCallback {
   }
 
   @Override
-  @Action
   protected void reportPropsChanged(@Nullable final JsPropertyMap<Object> nextProps) {
   }