+++ processor/src/test/resources/expected/com/example/arez/React4j_ExplicitlyEnabledCallback.java
@@ -36,7 +36,6 @@ abstract class React4j_ExplicitlyEnabledCallback extends ExplicitlyEnabledCallba
   }
 
   @Override
-  @Action
   protected void reportPropsChanged(@Nullable final JsPropertyMap<Object> nextProps) {
   }