+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.82](https://github.com/react4j/react4j/tree/v0.82) (2018-07-10)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.81...v0.82)