+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * **\[arez\]** Replace the jetbrains artifact with a J2CL compatible variant.
 * **\[arez\]** Made the `ReactArezEnvironment.ReactReactionEnvironment` class final so that Arez can
   change the `ReactionEnvironment` interface into a `@JsFunction` in the future to eliminate some overhead.
+* **\[processor\]** Enhance the generated subclasses of `react4j.arez.ReactArezComponent` to eliminate
+  the `arez.annotations.Action` annotation on the `reportPropsChanged(...)` method if there is no props
+  as the `@Action` annotation is not needed and will generate an invariant failure in Arez version `0.97+`.
 
 ### [v0.82](https://github.com/react4j/react4j/tree/v0.82) (2018-07-10)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.81...v0.82)