+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.83](https://github.com/react4j/react4j/tree/v0.83) (2018-07-16)
+[Full Changelog](https://github.com/react4j/react4j/compare/v0.82...v0.83)
 
 * **\[arez\]** Replace the jetbrains artifact with a J2CL compatible variant.
 * **\[arez\]** Made the `ReactArezEnvironment.ReactReactionEnvironment` class final so that Arez can